+++ aws/network/security_groups_app.tf
@@ -153,6 +153,16 @@ resource "aws_security_group_rule" "privatelink_api_ecs_ingress" {
   source_security_group_id = aws_security_group.api_ecs.id
 }
 
+resource "aws_security_group_rule" "privatelink_glue_job_ingress" {
+  description              = "Security group rule for Glue job ingress"
+  type                     = "ingress"
+  from_port                = 443
+  to_port                  = 443
+  protocol                 = "tcp"
+  security_group_id        = aws_security_group.privatelink.id
+  source_security_group_id = aws_security_group.glue_job.id
+}
+
 # Allow traffic from the app and from the lambdas
 resource "aws_security_group" "forms_database" {
   name        = "forms-database"